+++ NEWS.md
@@ -116,9 +116,12 @@ Features:
   the desired file.
 * Binary files are now displayed as a hex dump with ASCII
   representation (where applicable).
-* Added a `log_msg_line()` function that will return the
+* Added a `log_msg_line()` SQL function that will return the
   line number of the start of the currently focused
   message in the log view.
+* Added a `log_msg_values` column to the `all_logs` SQL
+  table that contains a JSON object with the top 5 values
+  for the fields extracted from the log message.
 
 Bug Fixes:
 * Binary data piped into stdin should now be treated the same